refactor character in nexus to use character factory
